fix: Fix race in scheduler (#25)
This fixes a data race in the scheduler that caused some child resources to be missed.

The implementation is now correct, but it could be made more efficient. The point is that we have to submit the sentinel indicating the start of the resolver before submitting the job, otherwise the job might send a "finished" signal before the "start" signal is sent, causing the counts to be equal before all resources have actually been processed.